cleanup(core): move storage classes to storage-api
